tidy up test_constraints.py
